dpif-netdev: e2e: Calculate ct_counters based on pointers values
In e2e_cache_calc_counters() calculate ct_counter of merged flow
a minimum value of pointers to MT flows objects. Guarantees uniqueness
and eliminates hash calculation. Propagate uintptr_t value to
offload layer.

Issue: 2333202

+++ lib/netdev-offload-dpdk.c
@@ -2370,7 +2370,7 @@ struct act_vars {
     struct flow_tnl tnl_mask;
     bool is_e2e_cache;
     uint32_t app_flows_counter;
-    uint32_t app_ct_counter;
+    uintptr_t app_ct_counter;
 };
 
 static struct rte_flow *
